// InSite/HB.InSite.Web/ClientApp/app/core/models/location.ts
export class Location {
    id: number;
    name: string;
    nameFormatted: string;
    value: string;
    cityId: number;
    stateId: number;
    metroId: number;
    countryId: number;
    regionId: number;
    cityName: string;
    stateName: string;
    metroName: string;
    countryName: string;
    regionName: string;
    createdBy: string;
    createdOn: Date;
    modifiedBy: string;
    modifiedOn: Date;
}