// InSite/HB.InSite.Web/ClientApp/app/core/models/propertyChain.ts
export class PropertyChain {
    brandImageURL: string;
    emailInvoice: boolean;
    faxInvoice: boolean;
    nonHotelChain: boolean;
    propertyChainID: number;
    propertyChainName: string;
    propertyChainScaleID: number;
    propertyChainTypeID: number;
    propertyParentChainID: number;
    createdBy: string;
    createdOn: Date;
    modifiedBy: string;
    modifiedOn: Date;
}